pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "simverse"
version = "0.1.0"
description = "A robotics simulation platform for training AI models — built on MuJoCo and Gymnasium"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "mujoco>=3.1.0",
    "gymnasium>=1.0.0",
    "numpy>=1.24.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
training = [
    "stable-baselines3>=2.3.0",
    "tensorboard>=2.15.0",
    "torch>=2.1.0",
    "shimmy>=2.0.0",
]
api = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.9",
]
cloud = [
    "mujoco-mjx>=3.1.0",
    "jax[cuda12]>=0.4.20",
]
viewer = [
    "mediapy>=1.2.0",
    "Pillow>=10.0.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]
all = ["simverse[training,api,viewer,dev]"]

[project.scripts]
simverse-train = "simverse.training.cli:main"
simverse-serve = "simverse.api.app:run_server"

[tool.hatch.build.targets.wheel]
packages = ["src/simverse"]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
